+++ user_management/src/user_management/utils.py
@@ -3,9 +3,12 @@ import string
 from io import BytesIO
 
 import requests
+from PIL import Image
 from django.core.files import File
 from django.core.files.base import ContentFile
-from PIL import Image
+
+from user.models import User
+from user_management import settings
 
 
 def generate_random_string(length):
@@ -30,3 +33,63 @@ def download_image_from_url(url, model_instance):
         return True
     else:
         return False
+
+
+def is_valid_username(username):
+    if username is None or username == '':
+        return False, 'Username empty'
+    if len(username) < settings.USERNAME_MIN_LENGTH:
+        return False, f'Username length {len(username)} < {settings.USERNAME_MIN_LENGTH}'
+    if len(username) > settings.USERNAME_MAX_LENGTH:
+        return False, f'Username length {len(username)} > {settings.USERNAME_MAX_LENGTH}'
+    if not username.isalnum():
+        return False, 'Username must be alphanumeric'
+    users = User.objects.filter(username=username)
+    if users.exists():
+        return False, f'Username {username} already taken'
+    return True, None
+
+
+def is_valid_email(email):
+    if email is None or email == '':
+        return False, 'Email empty'
+    users = User.objects.filter(email=email)
+    if users.exists():
+        return False, f'Email {email} already taken'
+    if len(email) > settings.EMAIL_MAX_LENGTH:
+        return False, f'Email length {len(email)} > {settings.EMAIL_MAX_LENGTH}'
+    if any(char in '!#$%^&*()=' for char in email):
+        return False, 'Invalid character in email address'
+    if '@' not in email:
+        return False, 'Email missing @'
+    if '.' not in email:
+        return False, 'Email missing "." character'
+    if email.count('@') > 1:
+        return False, 'Email contains more than one @ character'
+    local_part, domain_and_tld = email.rsplit('@', 1)
+    if len(local_part) < settings.EMAIL_LOCAL_PART_MIN_LENGTH:
+        return False, f'Local part length {len(local_part)} < {settings.EMAIL_LOCAL_PART_MIN_LENGTH}'
+    if domain_and_tld.count('.') == 0:
+        return False, 'Email missing TLD'
+    tld = domain_and_tld.rsplit('.')[-1]
+    if len(tld) > settings.TLD_MAX_LENGTH:
+        return False, f'TLD length {len(tld)} > {settings.TLD_MAX_LENGTH}'
+    return True, None
+
+
+def is_valid_password(password):
+    if password is None or password == '':
+        return False, 'Password empty'
+    if len(password) < settings.PASSWORD_MIN_LENGTH:
+        return False, f'Password length {len(password)} < {settings.PASSWORD_MIN_LENGTH}'
+    if len(password) > settings.PASSWORD_MAX_LENGTH:
+        return False, f'Password length {len(password)} > {settings.PASSWORD_MAX_LENGTH}'
+    if not any(char.isupper() for char in password):
+        return False, 'Password missing uppercase character'
+    if not any(char.islower() for char in password):
+        return False, 'Password missing lowercase character'
+    if not any(char.isdigit() for char in password):
+        return False, 'Password missing digit'
+    if not any(char in '!@#$%^&*()-_+=' for char in password):
+        return False, 'Password missing special character'
+    return True, None
